Handled reservation part using modal
