+++ src/components/settings/termsAndConditions/term&conditions.css
@@ -4,3 +4,7 @@
 #settings .terms-line {
   width: 94%;
 }
+
+#settings .terms-main {
+  background-color: white;
+}
