+++ src/components/villa/villa.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import {Link, Route, Switch, BrowserRouter as Router} from "react-router-dom";
-import VillaProfile from './villaProfile';
+import VillaProfile from './villaProfile/villaProfile';
 import SlideShow from '../slideShow/sildeShow';
+import Reserve from './reservation/reserve';
 class Villa extends Component {
     state = {  }
     render() { 
@@ -11,11 +12,12 @@ class Villa extends Component {
                     <Route exact path="/villa/villaProfile/">
                         <VillaProfile/>
                     </Route>
-                    <Router>
-                        <Route path="/villa/villaProfile/villaGallery/">
-                            <SlideShow/>
-                        </Route>
-                    </Router>
+                    <Route path="/villa/villaProfile/villaGallery/">
+                        <SlideShow/>
+                    </Route>
+                    <Route path="/villa/reserver/">
+                        <Reserve/>
+                    </Route>
                 </Switch>
             </Router>
          );
