+++ src/components/villa/villaProfile/villaProfile.jsx
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import './villaProfile.css';
-import sampleImage1 from '../villa/img/1.jpg';
-import sampleImage2 from '../villa/img/2.jpg';
-import sampleImage3 from '../villa/img/3.jpg';
-import sampleImage4 from '../villa/img/4.jpg';
-import sampleProfileImg from '../../assets/img/default-profile-picture.jpg';
-import plusImg from '../../assets/img/plus.png';
-import minusImg from '../../assets/img/minus.png';
+import sampleImage1 from '../img/1.jpg';
+import sampleImage2 from '../img/2.jpg';
+import sampleImage3 from '../img/3.jpg';
+import sampleImage4 from '../img/4.jpg';
+import sampleProfileImg from '../../../assets/img/default-profile-picture.jpg';
+import plusImg from '../../../assets/img/plus.png';
+import minusImg from '../../../assets/img/minus.png';
 import { Link } from "react-router-dom";
 import 'date-fns';
 import Grid from '@material-ui/core/Grid';
@@ -35,72 +35,10 @@ class VillaProfile extends Component {
             checkIn: new Date().toLocaleString(),
             checkOut: new Date().toLocaleString(),
             minCheckOut: '04/25/2021',
-            passangers: 1,
         }
     }
 
-    handleCheckInChange = (e) => {
-        if (this.state.checkOut > this.state.checkIn){
-            // To calculate the time difference of two dates
-            console.log("check in1 : "+e)
-            let Difference_In_Time = this.state.checkOut.getTime() - e.getTime();
-
-            // To calculate the no. of days between two dates
-            let Difference_In_Days = Difference_In_Time / (1000 * 3600 * 24);
-            console.log( Difference_In_Days )
-            this.setState({
-                checkIn: e,
-                minCheckOut: e,
-                stayingDays: Difference_In_Days,
-            })
-            return;
-        }
-        this.setState({
-            checkIn: e,
-            minCheckOut: e,
-            checkOut: e,
-        })
-        console.log(e)
-    }
-
-    handleCheckOutChange = (e) => {
-        // To calculate the time difference of two dates
-        console.log("check in2 : "+this.state.checkIn)
-        let Difference_In_Time = e.getTime() - this.state.checkIn.getTime();
-        
-        // To calculate the no. of days between two dates
-        let Difference_In_Days = Difference_In_Time / (1000 * 3600 * 24);
-        console.log( Difference_In_Days )
-        this.setState({
-            checkOut: e.value,
-            stayingDays: Difference_In_Days,
-        })
-
-    }
-
-    calculateDaysBetweenDates = (checkIn,checkOut) =>{
-        // To calculate the time difference of two dates
-        let Difference_In_Time = checkOut.getTime() - checkIn.getTime();
-        
-        // To calculate the no. of days between two dates
-        let Difference_In_Days = Difference_In_Time / (1000 * 3600 * 24);
-        return Difference_In_Days;
-    }
-
-    handleCounter = (select,operator) =>{
-        if (select === 1 && operator === "+" && this.state.passangers < this.state.placeMaxCapacity){
-            this.setState({
-                passangers: this.state.passangers + 1
-            })
-            return;
-        }
-        if (select === 1 && operator === "-" && this.state.passangers > 1){
-            this.setState({
-                passangers: this.state.passangers - 1
-            })
-            return;
-        }
-    }   
+ 
     render() { 
         return ( 
 
@@ -276,72 +214,10 @@ class VillaProfile extends Component {
                     </div>
 
                     <div className="villaProfile-reservation row mt-4">
-                        <div className="villaProfile-title ml-4">
-                            <h5>Reservation</h5>
-                            <div className="villaProfile-subtitle"></div>
-                        </div>
-                        <div className="col-xl-12 ml-4">
-                            <MuiPickersUtilsProvider utils={DateFnsUtils}>
-                                <div className="row">
-                                    <div className="col-xl-6">
-                                        <KeyboardDatePicker
-                                        width={50}
-                                        disablePast
-                                        margin="normal"
-                                        id="date-picker-dialog"
-                                        label="Select the arrival date"
-                                        format="MM/dd/yyyy"
-                                        value={this.state.checkIn}
-                                        onChange={this.handleCheckInChange}
-                                        minDateMessage=""
-                                        KeyboardButtonProps={{
-                                            'aria-label': 'change date',
-                                        }}
-                                        />
-                                    </div>
-
-                                    <div className="col-xl-6">
-                                        <KeyboardDatePicker
-                                        width={50}
-                                        disablePast
-                                        margin="normal"
-                                        id="date-picker-dialog"
-                                        label="Select the departure date"
-                                        minDate={this.state.minCheckOut}
-                                        // openTo={this.state.checkIn}
-                                        format="MM/dd/yyyy"
-                                        value={this.state.checkOut}
-                                        onChange={this.handleCheckOutChange}
-                                        minDateMessage=""
-                                        strictCompareDates={true}
-                                        KeyboardButtonProps={{
-                                            'aria-label': 'change date',
-                                        }}
-                                        />
-                                    </div>
-                                </div>
-                            </MuiPickersUtilsProvider>
-                        </div>
-                        <div className="villaProfile-counter ml-4 col-xl-12">
-                            <div className="row">
-                                <div className="col-xl-12">
-                                    <div className="row">
-                                        <div className="col-xl-6 mt-2">                  
-                                            <label className="mt-2" htmlFor="villaProfile-counter">number of passengers</label>
-                                            <div className="villaProfile-option">
-                                                <div className="d-flex flex-row villaProfile-counter justify-content-start w-100">
-                                                    <img name="passangers" onClick={(select, operator)=>this.handleCounter(1,"-")} className="" alt="minus icon" src={minusImg} />
-                                                    <div data-testid="villaProfile-passangers" className="villaProfile-counter-number"><span className="profileVilla-numOfPassanger">{this.state.passangers}</span></div>
-                                                    <img name="passangers" onClick={(select, operator)=>this.handleCounter(1,"+")} className="" alt="plus icon" src={plusImg} />
-                                                </div>
-                                            </div> 
-                                        </div>
-                                        <div className="col-xl-6 mt-4 villaProfile-reserveButton">
-                                            <button className="btn btn-primary">Reserve</button>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
+                        <div className="col-xl-6 mt-4 villaProfile-reserveButton">
+                            <Link to="/villa/reserver/">
+                                <button className="btn btn-primary">Reserve</button>  
+                            </Link>
                         </div>
                     </div>
                 </div>
